backend/requirements.txt
fastapi
starlette
boto3
python-dotenv
uvicorn
